+++ packages/terra-section-header/tests/nightwatch/OnClickSectionHeader.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import SectionHeader from '../../lib/SectionHeader';
+
+const OnClickSectionHeader = () =>
+  <SectionHeader
+    title="OnClick Section Header"
+    // eslint-disable-next-line no-alert
+    onClick={() => { window.alert('The accordion has been clicked!'); }}
+    className="onClickHeader"
+  />;
+
+export default OnClickSectionHeader;